updates verifyOTP

+++ controllers/authController.js
@@ -179,7 +179,7 @@ exports.Login =CatchAsync(async (req, res, next) => {
 
 //send OTP
 const sendOTPVerification = CatchAsync(async (user, res, next) => {
-  const {_id, email} = user;
+  const {_id} = user;
   const otp = `${Math.floor(10000 + Math.random() * 90000)}`;
   const hashedOTP = await bcrypt.hash(otp, 10);
 
@@ -200,13 +200,18 @@ exports.verifyOTP = CatchAsync(async (req, res, next) => {
   if(!email || !otp)
     return next(new AppError('Empty OTP details', 400));
   
-  const user = User.findOne({email});
+  const user = await User.findOne({email});
+  if (!user) 
+    return next(new AppError("User not found", 404));
+  
   const userId = user._id;
   const userOTPRecord = await UserOTPVerification.find({userId});
+
   if(userOTPRecord.length === 0)
     return next(new AppError("Account doesn't exist or has been verified already. Please signup or login"), 400);
 
   const {expireAt} = userOTPRecord[0];
+  
   if(expireAt < Date.now()){
     await UserOTPVerification.deleteMany({userId});
     return next(new AppError("OTP expired. Please request again.", 400));
